+++ client-sdk/src/errors.ts
@@ -1,0 +1,6 @@
+export class RTAPError extends Error {
+  constructor(message: string, options?: ErrorOptions) {
+    super(message, options);
+    this.name = 'RTAPError';
+  }
+}