+++ client-sdk/src/index.ts
@@ -1,0 +1,7 @@
+export * from './player';
+export * from './annotation';
+export * from './websocket';
+export * from './types';
+export * from './errors';
+
+export { RTAPClient } from './client';